[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -2,13 +2,13 @@
 
 using namespace std;
 
-int newOperator(const int a, const int b) {
+long long int newOperator(const long long int a, const long long int b) {
     return (a + b) * (a - b);
 }
 
 int main() {
 
-    int a, b;
+    long long int a, b;
     cin >> a >> b;
     cout << newOperator(a, b);
